HWKBTM-315 : Fix AngularUI tags for version < 0.14.0
AngularUI version >= 0.14.0 has prefixed their directives with uib-*,
but we are still using 0.13.x so these prefixes do not apply.
Fixing by removing the prefix, but probably to be reverted in the future
when both hawkular and bawkular-btm update to 0.14.x/1.x.

Also removed bootstrap libs incorrectly present in index.html.