HWKBTM-382 : Go to BTx Info on click @ Fault Count
When clicking on Fault Count chart for a specific BTxn, it wasn't
(immediately) leading the user to info page, which only occurred when
a digest cycle happened (could take several seconds). This fixes it by
wrapping it in a $timeout call.

Also took the chance to refactor it to a single function for both charts