HWKBTM-414 Fix Pan/Zoom resetting at each refresh
It seems D3 and the directive go out of sync, so the scale is applied
at element draw time and then again at rendering.

The workaround is to remove the applied transformations before rendering
so that it is rendered with default scale and re-apply once rendered.

Also disabled individual node dragging by default, which can be enabled
by using the node-drag attribute on the directive. It suffers from the
same issue, but solving it is a bit more complex, so leaving it out for
now.